Fixed duplicated code and added BaseConnection
